+++ server/src/main/java/seb40main026/mainproject/member/service/MemberServiceImpl.java
@@ -1,4 +1,4 @@
-package seb40main026.mainproject.member.unit;
+package seb40main026.mainproject.member.service;
 
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
